requirements.txt
# Requirements for dev environment
wheel>=0.22
django-modeltranslation>=0.12
wagtail>=2.0
pytest
pytest-django>=4.5.2
pytest-pythonpath
